+++ src/app/models/pkmnModels/pkmnShape.model.ts
@@ -1,4 +1,4 @@
-import { Name, NamedLink } from "./core.model";
+import { Name, NamedLink } from "../core.model";
 
 /**
  * Represents the shape of a Pokémon.
